fix: add functionality for inner folders copy to task 04-copy-directory

+++ 04-copy-directory/index.js
@@ -4,39 +4,46 @@ const path = require("path");
 const pathToOrigFolder = path.join(__dirname, "files");
 const pathToCopiedFolder = path.join(__dirname, "files-copy");
 
-async function copyFolderContent() {
-  await fsPromisesModule.mkdir(pathToCopiedFolder, { recursive: true });
+async function copyFolder(origFolder, finalFolder) {
+  await fsPromisesModule.mkdir(finalFolder, { recursive: true });
 
-  const folderContent = await fsPromisesModule.readdir(pathToOrigFolder, {
+  const folderContent = await fsPromisesModule.readdir(origFolder, {
     withFileTypes: true,
   });
 
   const arrayOfFileNames = [];
 
   for (const contentItem of folderContent) {
-    const fileName = contentItem.name;
+    if (contentItem.isFile()) {
+      const fileName = contentItem.name;
 
-    const pathToOrigFile = path.join(pathToOrigFolder, fileName);
-    const pathToCopiedFile = path.join(pathToCopiedFolder, fileName);
-    fsPromisesModule.copyFile(pathToOrigFile, pathToCopiedFile);
+      const pathToOrigFile = path.join(origFolder, fileName);
+      const pathToCopiedFile = path.join(finalFolder, fileName);
+      fsPromisesModule.copyFile(pathToOrigFile, pathToCopiedFile);
 
-    arrayOfFileNames.push(fileName);
+      arrayOfFileNames.push(fileName);
+    } else if (contentItem.isDirectory()) {
+      const folderName = contentItem.name;
+      const pathToOrigFolder = path.join(origFolder, folderName);
+      const pathToCopiedFolder = path.join(finalFolder, folderName);
+      copyFolder(pathToOrigFolder, pathToCopiedFolder);
+    }
   }
 
-  const copiedFolderContent = await fsPromisesModule.readdir(
-    pathToCopiedFolder,
-    {
-      withFileTypes: true,
-    }
-  );
+  const copiedFolderContent = await fsPromisesModule.readdir(finalFolder, {
+    withFileTypes: true,
+  });
 
   for (const contentItem of copiedFolderContent) {
-    const fileName = contentItem.name;
-    const pathToFile = path.join(pathToCopiedFolder, fileName);
-    if (!arrayOfFileNames.includes(fileName)) {
-      fsPromisesModule.rm(pathToFile);
+    if (contentItem.isFile()) {
+      const fileName = contentItem.name;
+      const pathToFile = path.join(finalFolder, fileName);
+      if (!arrayOfFileNames.includes(fileName)) {
+        fsPromisesModule.rm(pathToFile);
+      }
     }
   }
 }
 
-copyFolderContent();
+copyFolder(pathToOrigFolder, pathToCopiedFolder);
+
